backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pdfplumber==0.10.2
pymupdf==1.23.7
python-docx==1.0.1
requests==2.31.0
arxiv==2.0.0
biopython==1.81
python-dotenv==1.0.0
groq==0.4.1